requirements.txt
streamlit
langchain
langchain-google-genai
langchain-tavily
python-dotenv
pydantic<2